Refactor shared logic, inject page revealers
