+++ packages/client-sdk/src/widgets/attribute.ts
@@ -1,0 +1,6 @@
+/**
+ * Convert enums to web component attribute style.
+ */
+export default function toAttributeStyle(attribute: string): string {
+  return attribute.replace(/_/g, "-");
+}
